+++ packages/spacecat-shared-google-client/CHANGELOG.md
@@ -1,3 +1,17 @@
+# [@adobe/spacecat-shared-google-client-v1.1.7](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.1.6...@adobe/spacecat-shared-google-client-v1.1.7) (2024-06-07)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/fetch to v4.1.8 ([#255](https://github.com/adobe/spacecat-shared/issues/255)) ([ef89237](https://github.com/adobe/spacecat-shared/commit/ef8923738d9e0591693e47f4bd4345651d180389))
+
+# [@adobe/spacecat-shared-google-client-v1.1.6](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.1.5...@adobe/spacecat-shared-google-client-v1.1.6) (2024-06-06)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/fetch to v4.1.4 ([#253](https://github.com/adobe/spacecat-shared/issues/253)) ([687d739](https://github.com/adobe/spacecat-shared/commit/687d73947f15344ed6f4d6e74f223aa838ec3d6a))
+
 # [@adobe/spacecat-shared-google-client-v1.1.5](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-google-client-v1.1.4...@adobe/spacecat-shared-google-client-v1.1.5) (2024-06-01)