+++ packages/spacecat-shared-rum-api-client/CHANGELOG.md
@@ -1,3 +1,30 @@
+# [@adobe/spacecat-shared-rum-api-client-v2.0.2](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-rum-api-client-v2.0.1...@adobe/spacecat-shared-rum-api-client-v2.0.2) (2024-06-07)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/fetch to v4.1.8 ([#255](https://github.com/adobe/spacecat-shared/issues/255)) ([ef89237](https://github.com/adobe/spacecat-shared/commit/ef8923738d9e0591693e47f4bd4345651d180389))
+
+# [@adobe/spacecat-shared-rum-api-client-v2.0.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-rum-api-client-v2.0.0...@adobe/spacecat-shared-rum-api-client-v2.0.1) (2024-06-06)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/fetch to v4.1.4 ([#253](https://github.com/adobe/spacecat-shared/issues/253)) ([687d739](https://github.com/adobe/spacecat-shared/commit/687d73947f15344ed6f4d6e74f223aa838ec3d6a))
+
+# [@adobe/spacecat-shared-rum-api-client-v2.0.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-rum-api-client-v1.8.4...@adobe/spacecat-shared-rum-api-client-v2.0.0) (2024-06-06)
+
+
+### Features
+
+* client for rum bundler ([#252](https://github.com/adobe/spacecat-shared/issues/252)) ([854fc58](https://github.com/adobe/spacecat-shared/commit/854fc583d7184048d22c357cf11f349d3c1cfc9d))
+
+
+### BREAKING CHANGES
+
+* With this change RUMAPIClient will not be using
+helix-run-query as datasource anymore
+
 # [@adobe/spacecat-shared-rum-api-client-v1.8.4](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-rum-api-client-v1.8.3...@adobe/spacecat-shared-rum-api-client-v1.8.4) (2024-06-01)